test(logic): split combat_resolver part1 tests for file-size gate (Refs #2288)
Move deployment limits, general medal persistence, and medal-cap cases into
combat_resolver_test_part1_limits_test.dart; keep core engagement and
resolveBattleContext coverage in part1. Each file stays under the 400-line
logic test gate.
